+++ changelog.md
@@ -37,6 +37,7 @@ Codion Change Log
 - EntityConditions.conditionModel() renamed condition().
 - EntityConditionModel bug fixed, get(Column<T> column) reintroduced, required for type inference.
 - EntityEditModel.EditorValue.original() added.
+- EntityExportModel refactored from EntityTableExportModel, related refactoring.
 ### is.codion.swing.framework.model
 - SwingEntityConditionModelFactory renamed SwingEntityConditions.
 ### is.codion.swing.framework.ui